+++ app/demo/package.json
@@ -14,7 +14,8 @@
     "chokidar": "^3.5.3",
     "esbuild": "^0.19.4",
     "live-server": "^1.2.2",
-    "luma-js": "file:../../package",
+    "luma-js": "file:../../packages/luma-js",
+    "luma-router": "file:../../packages/luma-router",
     "snabbdom": "^3.5.1",
     "snabbdom-to-html": "^7.1.0"
   }
